Finally set up prettier.
I am changing the formatting from hard tabs of 4 width to spaces of 2 width. I usually go with tabs on codebases, but spaces are used quite commonly in the code I work with on a day-to-day basis, so I would rather just not have to deal with the inconsistencies.

That being said, I am not a fan of 80col. This is not the 1990s, we have big monitors now. We can survive with longer lines.