Organize images into respective publication year/month date folders
Images were disorganized as Hugo did it based on when the file was uploaded (with the Ghost site going up in March of 2022), not publication of the article (which was backdated since the content was lifted from my stuff on Solus). Fixed that.

Renamed a bunch of the featured images to actually be related to their release.

Added Alert in and use for Wayland article.